recipes-tisdk: ti-tisdk-makefile: Rules.make: Allow overwriting DESTDIR
DESTDIR in the Rules.make points to the path where
the files should be installed.

setup scripts update this to the targetNFS location but
it cannot be used for installing files in the SD card.

When DESTDIR is defined conditionally, it allows following
DESTDIR=/path/to/sd/card make install

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>