recipes-demos: oob-demo: Add service to update weston wallpaper
Add system service to update the wallpaper based on the board name from
the device-tree compatible string. The service grep for the EAIK board
name in the compatible string and update the weston.ini background-image
parameter pointing to the corresponding wallpaper. The service executes
before the weston service is invoked.

Signed-off-by: Sinthu Raja <sinthu.raja@ti.com>