recipes-demo: wifi-oob: Add a new recipe to enable WiFi access point
J7 SK platform uses the intel 9260 wifi module. To demonstrate
and validate the same, add support for wifi oob demo for J7 SK.
This patch adds a systemd service that invokes the demo to
configure the intel 9260 wifi as an access point during
the board boot-up.

The demo includes the following changes
* Install a systemd service that invokes the demo when the board boots.
* Add conf and script files to load the wifi modules, configure
and enable the intel 9260 wifi interface as an access point.

Signed-off-by: Dasnavis Sabiya <sabiya.d@mistralsolutions.com>