linux-ti-staging: Add patch fix for LCD issue on AM57x IDK
Issue: u-boot fails to load the dtb because the LCD DTB is
missing in the boot partition with the LCD connected to the IDK
board.

Rootcause: While porting to 5.10 the LCD DTBs are converted to DTB
overlay and missed to merge the overlays into single DTB

Fix: Merge all the DTB overlays into single DTB for both AM571x and
AM572x IDK

JIRA-ID: LCPD-28152- AM57x-idk: Failed to load '/boot/am571x-idk-lcd-
osd101t2587.dtb.

Signed-off-by: Sinthu Raja <sinthu.raja@ti.com>